+++ Demo/src/main/java/com/braintreepayments/demo/DemoActivity.java
@@ -16,16 +16,12 @@ import androidx.appcompat.widget.Toolbar;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
 import androidx.navigation.NavController;
 import androidx.navigation.fragment.NavHostFragment;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
 import com.braintreepayments.api.BraintreeClient;
-import com.braintreepayments.api.BraintreeClientFactory;
-import com.braintreepayments.api.BrowserSwitchResult;
-import com.braintreepayments.api.DemoAuthorizationProvider;
 
 import java.util.Arrays;
 import java.util.List;